[NP-1917] Fix initial metrics retrieval.
If the initial metrics retrieval failed (which might often happen
because metrics-collector fails to query Prometheus because Prometheus
is not ready yet), the collector loop would not start. Fix contains two
changes:

- If initial stats gathering fails, we start the collection loop anyway,
so after a minute or so we should have stats
- If the metrics cache is queried because we're trying to calculate a
score (because we're trying to deploy an application) and it's empty
(because the initial gathering failed), run the gather call at that very
moment.